% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2odyssey.R
\name{create_java_call}
\alias{create_java_call}
\title{Create a Java call to start the h2o worker}
\usage{
create_java_call(network = NULL, memory = 2, path = NULL)
}
\arguments{
\item{network}{Network information of the SLURM job}

\item{memory}{The amount of memory per node in GB}

\item{path}{Path of h2o.jar}
}
\value{
A vector of strings to start the h2o workers
}
\description{
Create a Java call to start the h2o worker
}
\examples{
node_list <- detect_nodes()
ips <- get_ips(node_list)
network <- get_network(ips)
args <- create_java_call(network)

}
