% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2odyssey.R
\name{detect_nodes}
\alias{detect_nodes}
\title{Detect the nodes of a SLURM job}
\usage{
detect_nodes()
}
\arguments{
\item{None}{}
}
\value{
A list of nodes
}
\description{
Detect the nodes of a SLURM job
}
\examples{
detect_nodes()

}
