% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2odyssey.R
\name{get_ips}
\alias{get_ips}
\title{Detect the IPs of the nodes of a SLURM job}
\usage{
get_ips(node_list = NULL)
}
\arguments{
\item{node_list}{Nodes of the SLURM jobs}
}
\value{
List of SLURM IPs
}
\description{
Detect the IPs of the nodes of a SLURM job
}
\examples{
node_list <- detect_nodes()
ips <- get_ips(node_list)

}
