% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2odyssey.R
\name{get_network}
\alias{get_network}
\title{Detect the network of the nodes of a SLURM job}
\usage{
get_network(ips = NULL)
}
\arguments{
\item{ips}{IPs of the nodes of the SLURM job}
}
\value{
List of SLURM network
}
\description{
Detect the network of the nodes of a SLURM job
}
\examples{
node_list <- detect_nodes()
ips <- get_ips(node_list)
network <- get_network(ips)

}
