% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2odyssey.R
\name{make_nodes_known_hosts}
\alias{make_nodes_known_hosts}
\title{Add the SLURM nodes to the list of known hosts}
\usage{
make_nodes_known_hosts(ips = NULL)
}
\arguments{
\item{ips}{IPs of the nodes of the SLURM job}
}
\value{
None This function is used for its side effects.
}
\description{
Add the SLURM nodes to the list of known hosts
}
\examples{
node_list <- detect_nodes()
ips <- get_ips(node_list)
make_nodes_known_hosts(ips)
}
