% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2odyssey.R
\name{start_h2o_cluster}
\alias{start_h2o_cluster}
\title{Start an h2o cluster with an h2o worker on each node of the SLURM cluster}
\usage{
start_h2o_cluster(memory = 2,
  path = "/n/home03/cchoirat/apps/R/h2o/java/h2o.jar")
}
\arguments{
\item{memory}{The amount of memory per node in GB}

\item{path}{Path of h2o.jar}
}
\value{
None This function is used for its side effects.
}
\description{
Start an h2o cluster with an h2o worker on each node of the SLURM cluster
}
\examples{
start_h2o_cluster()
library(h2o)
# Connect to our existing h2o cluster.
# Do not try to start a new server from R.
h2o.init(startH2O = FALSE)
# Code here...
h2o.shutdown()

}
