% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2odyssey.R
\name{start_h2o_cluster}
\alias{start_h2o_cluster}
\title{Start an h2o cluster with an h2o worker on each node of the SLURM cluster}
\usage{
start_h2o_cluster()
}
\arguments{
\item{None}{}
}
\value{
None This function is used for its side effects.
}
\description{
Start an h2o cluster with an h2o worker on each node of the SLURM cluster
}
\examples{
start_h2o_cluster()
library(h2o)
# Connect to our existing h2o cluster.
# Do not try to start a new server from R.
h2o.init(startH2O = FALSE)
# Code here...
h2o.shutdown()
}
