% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2odyssey.R
\name{start_h2o_workers}
\alias{start_h2o_workers}
\title{Start an h2o worker on each node of the SLURM cluster}
\usage{
start_h2o_workers(ips = NULL, memory = 2,
  path = "/n/home03/cchoirat/apps/R/h2o/java/h2o.jar")
}
\arguments{
\item{ips}{IPs of the nodes of the SLURM job}

\item{memory}{The amount of memory per node in GB}

\item{path}{Path of h2o.jar}
}
\value{
None This function is used for its side effects.
}
\description{
Start an h2o worker on each node of the SLURM cluster
}
\examples{
node_list <- detect_nodes()
ips <- get_ips(node_list)
start_h2o_workers(ips)

}
